fix: some cleanup and refactoring
