+++ components/ads-client/src/client.rs
@@ -98,18 +98,6 @@ where
         self.client.clear_cache()
     }
 
-    // Shutdown the db connection and drop references to telemetry callbacks.
-    // Should be used only when dropping the ads client, this may be extended to drop more things.
-    pub fn shutdown_client(&mut self) -> Result<(), rusqlite::Error> {
-        // Drop telemetry (within the telemetry wrapper)
-        self.telemetry.shutdown();
-
-        // Shutdown DB
-        self.client.shutdown_db()?;
-
-        Ok(())
-    }
-
     pub fn get_context_id(&self) -> context_id::ApiResult<String> {
         self.context_id_provider.context_id()
     }
@@ -286,7 +274,7 @@ pub enum ClientOperationEvent {
 
 #[cfg(test)]
 mod tests {
-    use std::{assert_eq, assert_ne, sync::Arc};
+    use std::assert_eq;
 
     use crate::{
         ffi::telemetry::MozAdsTelemetryWrapper,
@@ -534,74 +522,5 @@ mod tests {
 
         m1.assert();
         m2.assert();
-    }
-
-    #[test]
-    fn test_shutdown_telemetry_basic() {
-        viaduct_dev::init_backend_dev();
-
-        // test with client created from config
-        let noop_telemetry = MozAdsTelemetryWrapper::noop();
-        let weak_reference = Arc::downgrade(
-            &noop_telemetry
-                .clone_inner_arc()
-                .expect("Inner telemetry should be Some before dropping"),
-        );
-        let config = AdsClientConfig {
-            cache_config: None,
-            context_id_provider: None,
-            environment: Environment::Test,
-            telemetry: noop_telemetry,
-        };
-        let mut client = AdsClient::new(config);
-
-        // weak ref will show 0 strong references when the Arc<dyn MozAdsTelemetry> is gone.
-        assert_ne!(weak_reference.strong_count(), 0);
-        client.shutdown_client().unwrap();
-        assert_eq!(weak_reference.strong_count(), 0);
-
-        // test also with internal function from_mars
-        let noop_telemetry = MozAdsTelemetryWrapper::noop();
-        let weak_reference = Arc::downgrade(
-            &noop_telemetry
-                .clone_inner_arc()
-                .expect("Inner telemetry should be Some before dropping"),
-        );
-        let cache = HttpCache::builder("test_shutdown_telemetry")
-            .build()
-            .unwrap();
-        let mars_client = MARSClient::new(Environment::Test, Some(cache), noop_telemetry);
-        let mut client = new_with_mars_client(mars_client);
-
-        // weak ref will show 0 strong references when the Arc<dyn MozAdsTelemetry> is gone.
-        assert_ne!(weak_reference.strong_count(), 0);
-        client.shutdown_client().unwrap();
-        assert_eq!(weak_reference.strong_count(), 0);
-    }
-
-    #[test]
-    fn test_shutdown_is_idempotent() {
-        viaduct_dev::init_backend_dev();
-
-        let noop_telemetry = MozAdsTelemetryWrapper::noop();
-        let weak_reference = Arc::downgrade(
-            &noop_telemetry
-                .clone_inner_arc()
-                .expect("Inner telemetry should be Some before dropping"),
-        );
-        // A real cache so the second shutdown exercises the db close path.
-        let cache = HttpCache::builder("test_shutdown_is_idempotent")
-            .build()
-            .unwrap();
-        let mars_client = MARSClient::new(Environment::Test, Some(cache), noop_telemetry);
-        let mut client = new_with_mars_client(mars_client);
-
-        client.shutdown_client().unwrap();
-        assert_eq!(weak_reference.strong_count(), 0);
-
-        // Repeated shutdowns must not error or re-close an already closed connection.
-        client.shutdown_client().unwrap();
-        client.shutdown_client().unwrap();
-        assert_eq!(weak_reference.strong_count(), 0);
     }
 }
